+++ SuperMarioBros3/WorldMap.cpp
@@ -123,7 +123,7 @@ void CWorldMap::_ParseSection_ANIMATIONS(string line)
 	LPANIMATION ani = new CAnimation();
 
 	int ani_id = atoi(tokens[0].c_str());
-	for (int i = 1; i < tokens.size(); i += 2)	// why i+=2 ?  sprite_id | frame_time  
+	for (unsigned int i = 1; i < tokens.size(); i += 2)	// why i+=2 ?  sprite_id | frame_time  
 	{
 		int sprite_id = atoi(tokens[i].c_str());
 		int frame_time = atoi(tokens[i + 1].c_str());
@@ -144,7 +144,7 @@ void CWorldMap::_ParseSection_ANIMATION_SETS(string line)
 
 	CAnimations* animations = CAnimations::GetInstance();
 
-	for (int i = 1; i < tokens.size(); i++)
+	for (unsigned int i = 1; i < tokens.size(); i++)
 	{
 		int ani_id = atoi(tokens[i].c_str());
 
@@ -163,8 +163,8 @@ void CWorldMap::_ParseSection_OBJECTS(string line)
 	if (tokens.size() < 3) return; // skip invalid lines - an object set must have at least id, x, y
 
 	int object_type = atoi(tokens[0].c_str());
-	float x = atof(tokens[1].c_str());
-	float y = atof(tokens[2].c_str());
+	float x =(float) atof(tokens[1].c_str());
+	float y =(float) atof(tokens[2].c_str());
 
 	int ani_set_id = atoi(tokens[3].c_str());
 
@@ -298,14 +298,14 @@ void CWorldMap::SetCamera()
 	zone->GetZone(lZone, tZone, rZone, bZone);
 
 
-	CGame::GetInstance()->SetCamPos(round(lZone), round(tZone));
+	CGame::GetInstance()->SetCamPos((float)round(lZone), (float)round(tZone));
 }
 
 void CWorldMap::Render()
 {
 	if (map)
 		this->map->Render();
-	for (int i = 0; i < objects.size(); i++)
+	for (unsigned int i = 0; i < objects.size(); i++)
 	{
 		objects[i]->Render();
 	}
@@ -321,7 +321,7 @@ void CWorldMap::Unload()
 {
 	if(player)
 		CBackUp::GetInstance()->BackUpMarioWM(player);
-	for (int i = 0; i < objects.size(); i++)
+	for (unsigned int i = 0; i < objects.size(); i++)
 		delete objects[i];
 
 	objects.clear();
