+++ SuperMarioBros3/main.cpp
@@ -136,7 +136,7 @@ int Run()
 {
 	MSG msg;
 	int done = 0;
-	DWORD frameStart = GetTickCount64();
+	DWORD frameStart = (DWORD)GetTickCount64();
 	DWORD tickPerFrame = 1000 / MAX_FRAME_RATE;
 
 	while (!done)
